Changed the MediaInfo DurationMilliseconds to DurationInSeconds and made it decimal to combat any possible rounding issues. Other properties based on it also changed to decimal type where appropriate. Added a bunch of tooltips.

+++ apps/VideoConversionApp/ViewModels/Components/VideoPlayerViewModel.cs
@@ -41,9 +41,9 @@ public partial class VideoPlayerViewModel : ViewModelBase
     [ObservableProperty]
     public partial float VideoPlayerFov { get; set; }
     [ObservableProperty]
-    public partial double CropTimelineStartTime { get; set; }
+    public partial decimal CropTimelineStartTime { get; set; }
     [ObservableProperty]
-    public partial double CropTimelineEndTime { get; set; } 
+    public partial decimal CropTimelineEndTime { get; set; } 
     
     public VideoPlayerViewModel(ConvertibleVideoModel? sourceConvertibleVideo, KeyFrameVideo? video)
     {
